+++ demo/loading_phase_mechanic.gd
@@ -9,11 +9,13 @@ func _ready():
 	t.start(1)
 
 func _on_timeout():
-	if count == 2:
-		SceneManager.load_percent_changed.emit(130)
+	count += 1
+	if count == 1:
+		SceneManager.load_percent_changed.emit(80 + randi_range(0, 9))
+	elif count == 2:
+		SceneManager.load_percent_changed.emit(90 + randi_range(0, 9))
+	if count == 3:
+		SceneManager.load_percent_changed.emit(100)
 		SceneManager.load_finished.emit()
 		t.timeout.disconnect(_on_timeout)
-	else:
-		count += 1
-		SceneManager.load_percent_changed.emit(100 + (count * 10) + randi_range(0, 9))
-		t.start(count + 1)
+	t.start(count + 1)